+++ tests/Unit/Schema/Execution/Fixtures/FooContext.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Unit\Schema\Execution\Fixtures;
 
+use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Foundation\Auth\User;
 use Illuminate\Http\Request;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
@@ -10,10 +11,7 @@ final class FooContext implements GraphQLContext
 {
     public const FROM_FOO_CONTEXT = 'custom.context';
 
-    /**
-     * @var \Illuminate\Http\Request
-     */
-    private $request;
+    private Request $request;
 
     public function __construct(Request $request)
     {
@@ -23,6 +21,10 @@ final class FooContext implements GraphQLContext
     public function user(): User
     {
         return new User();
+    }
+
+    public function setUser(?Authenticatable $user): void
+    {
     }
 
     public function request(): Request