+++ tests/Utils/Validators/EmailCustomAttributeValidator.php
@@ -6,8 +6,6 @@ use Nuwave\Lighthouse\Validation\Validator;
 
 final class EmailCustomAttributeValidator extends Validator
 {
-    public const MESSAGE = 'The email address must be a valid email address.';
-
     public function rules(): array
     {
         return [